requirements.txt
requests
bs4
ruamel.yaml